requirements.txt
fastapi
uvicorn[standard]
torch
torchvision
easyocr
opencv-python-headless
pillow
numpy
psycopg2-binary
pandas
python-multipart